% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizedTCGAData-package.r
\docType{data}
\name{project_ids}
\alias{project_ids}
\title{project_ids
A named character vector: mapping sumbitter_id (can be seen as patient ID) to its TCGA project ID}
\format{An object of class \code{character} of length 14551.}
\usage{
project_ids
}
\description{
project_ids
A named character vector: mapping sumbitter_id (can be seen as patient ID) to its TCGA project ID
}
\keyword{datasets}
