% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizedTCGAData-package.r
\docType{data}
\name{surv.plot}
\alias{surv.plot}
\title{surv.plot}
\format{A data frame with four variables: \code{survivalEstimate}, \code{id},
  \code{censored}, and \code{time}}
\usage{
surv.plot
}
\description{
Patient survival information downloaded from https://portal.gdc.cancer.gov
For detailed information: see section "Survival analysis" in https://docs.gdc.cancer.gov/Data_Portal/PDF/Data_Portal_UG.pdf
}
\keyword{datasets}
