% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizedTCGAData-package.r
\docType{data}
\name{project_ids}
\alias{project_ids}
\title{project_ids}
\format{An object of class \code{character} of length 14551.}
\usage{
project_ids
}
\description{
A named character vector: mapping case_id (i.e., patient ID) to
the TCGA project ID they belong to
}
\examples{
library(ExperimentHub)
eh <- ExperimentHub()
myfiles <- query(eh, "HarmonizedTCGAData")
project_ids <- myfiles[[2]]
# project_ids <- myfiles[['EH1015']]
head(project_ids)
}
\keyword{datasets}
