% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarmonizedTCGAData-package.r
\docType{data}
\name{Wall}
\alias{Wall}
\title{Wall
Wall is a complex list and contains lists inside list.
Precisely, Wall is a list (five cancer types) of list (six feature normalization types: raw.all, raw.sel, log.all, log.sel, vst.sel, normalized) of list (three feature spaces or views: fpkm, mirna, and methy450) of matrices.
The rownames of each matrix is the submitter_id (can be seen as a patient id),
and the column names of each matrix is the aliquot ID (which contains the submitter_id as prefix).
Based on these aliquot IDs, users can download original data from https://portal.gdc.cancer.gov/repository.}
\format{An object of class \code{list} of length 5.}
\usage{
Wall
}
\description{
Wall
Wall is a complex list and contains lists inside list.
Precisely, Wall is a list (five cancer types) of list (six feature normalization types: raw.all, raw.sel, log.all, log.sel, vst.sel, normalized) of list (three feature spaces or views: fpkm, mirna, and methy450) of matrices.
The rownames of each matrix is the submitter_id (can be seen as a patient id),
and the column names of each matrix is the aliquot ID (which contains the submitter_id as prefix).
Based on these aliquot IDs, users can download original data from https://portal.gdc.cancer.gov/repository.
}
\keyword{datasets}
